[#69] move ActivityAt vo to common package
