+++ src/main/java/depromeet/lessonfour/server/activityrecord/app/service/QueryActivityRecordUseCase.java
@@ -7,9 +7,9 @@ import org.springframework.transaction.annotation.Transactional;
 import depromeet.lessonfour.server.activityrecord.app.dto.response.GetActivityRecordsResponse;
 import depromeet.lessonfour.server.activityrecord.domain.entity.ActivityRecord;
 import depromeet.lessonfour.server.activityrecord.domain.repository.ActivityRecordRepository;
-import depromeet.lessonfour.server.activityrecord.domain.vo.ActivityAt;
 import depromeet.lessonfour.server.common.annotation.UseCase;
 import depromeet.lessonfour.server.common.api.code.ErrorCode;
+import depromeet.lessonfour.server.common.domain.vo.ActivityAt;
 import depromeet.lessonfour.server.common.exception.ServerException;
 import lombok.RequiredArgsConstructor;
