+++ src/main/java/depromeet/lessonfour/server/activityrecord/infra/repository/ActivityRecordRepositoryImpl.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 
 import depromeet.lessonfour.server.activityrecord.domain.entity.ActivityRecord;
 import depromeet.lessonfour.server.activityrecord.domain.repository.ActivityRecordRepository;
-import depromeet.lessonfour.server.activityrecord.domain.vo.ActivityAt;
+import depromeet.lessonfour.server.common.domain.vo.ActivityAt;
 import lombok.RequiredArgsConstructor;
 
 @Repository
