+++ src/main/java/depromeet/lessonfour/server/activityrecord/infra/repository/QueryDslActivityRecordRepository.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
 import depromeet.lessonfour.server.activityrecord.domain.entity.ActivityRecord;
-import depromeet.lessonfour.server.activityrecord.domain.vo.ActivityAt;
+import depromeet.lessonfour.server.common.domain.vo.ActivityAt;
 import lombok.RequiredArgsConstructor;
 
 @Repository
